Aurec Project
Fixed Bits Per Sample Issue

+++ Aurec/src/common/aurec/writers/WavWriter.java
@@ -103,7 +103,7 @@ public class WavWriter extends TrackWriter {
 		header[31] = (byte) ((byteRate >> 24) & 0xff);
 		header[32] = (byte) (2 * 16 / 8); // block align
 		header[33] = 0;
-		header[34] = mTrack.getBitsPerSample(); // bits per sample
+		header[34] = (byte) mTrack.getBitsPerSample(); // bits per sample
 		header[35] = 0;
 		header[36] = 'd';
 		header[37] = 'a';
